# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Buddies API
  version: 1.0.0
  description: |
    Dual-sided B2B reverse-auction marketplace API (consumer restaurants + suppliers).
    Auth: `Authorization: Bearer <supabase_jwt>` or dev `Bearer dev:<role>:<userId>`.
servers:
  - url: http://localhost:8000/v1
...
paths:
...
  /consumer/bid-requests:
    get:
      tags: [consumer]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: List own bid requests }
    post:
      tags: [consumer]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [privacyAccepted, items]
              properties:
                privacyAccepted: { type: boolean }
                budgetPaise: { type: integer }
                durationHours: { type: integer }
                deliveryWindow: { type: string }
                lat: { type: number }
                lng: { type: number }
                items:
                  type: array
                  items:
                    type: object
                    required: [name, quantity, unit]
                    properties:
                      name: { type: string }
                      quantity: { type: number }
                      unit: { type: string }
                      productCategory: { type: string }
      responses:
        '201': { description: Bid request created }
  /consumer/bid-requests/{id}/reorder:
    post:
      tags: [consumer]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: Cloned auction }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...